feat: quantile loss instead of discriminator
get rid of GAN by using a quantile loss on the RT distributions.

works well for constant, linear and saturating non-linear drifts. dynamic dt is best, but even 100 steps seems to work well (although it can be improved)

did not work for exponential drift. approximates the distribution but the drift rate does not become exponentially increasing.